+++ src/tl/Gestor.h
@@ -21,7 +21,7 @@ public:
 
     void insertVertice(string nombre);
 
-    string insertaArista(string salida, string llegada, int peso);
+    string insertaArista(string salida, string llegada, float peso);
 
     string listaAdyacencia();
